* reverted tests * reverted notification methods * enable to provide custom executor to non-blocking * add support for returning Futures from submitting notifications
